// modules/apps/account/account-admin-web/src/main/resources/META-INF/resources/account_entries_admin/js/types.ts
export interface MultiSelectItem {
    label: string;
    value: string;
}

export interface ValidatableMultiSelectItem extends MultiSelectItem {
    errorMessage?: string;
}

export interface InputGroup {
    accountRoles: ValidatableMultiSelectItem[];
    emailAddresses: ValidatableMultiSelectItem[];
    id: string;
}